mod/fs: rewrite

+++ mod/fs/src/writer.go
@@ -4,13 +4,10 @@ import (
 	"crypto/rand"
 	"encoding/hex"
 	"errors"
-	"github.com/cryptopunkscc/astrald/astral"
-	"github.com/cryptopunkscc/astrald/mod/fs"
 	"github.com/cryptopunkscc/astrald/mod/objects"
 	"github.com/cryptopunkscc/astrald/object"
 	"os"
 	"path/filepath"
-	"strings"
 	"sync/atomic"
 )
 
@@ -73,47 +70,25 @@ func (w *Writer) Commit() (*object.ID, error) {
 	var newPath = filepath.Join(w.path, objectID.String())
 
 	stat, err := os.Stat(newPath)
-	if err != nil {
+	if err == nil && stat.Mode().IsRegular() {
+		// we already have this object
+		os.Remove(oldPath)
+	} else {
 		err = os.Rename(oldPath, newPath)
 		if err != nil {
 			os.Remove(oldPath)
 		}
-	} else {
-		// we already have this object
-		os.Remove(oldPath)
 	}
 
+	// make sure the path is accessible
 	stat, err = os.Stat(newPath)
-	if err != nil {
-		return objectID, err
+	if err != nil || !stat.Mode().IsRegular() {
+		return nil, err
 	}
 
-	err = w.repo.mod.db.Create(&dbLocalFile{
-		Path:    newPath,
-		DataID:  objectID,
-		ModTime: stat.ModTime(),
-	}).Error
+	w.repo.pushAdded(objectID)
 
-	switch {
-	case err == nil:
-		w.repo.mod.Objects.Receive(&fs.EventFileAdded{
-			Path:     astral.String16(newPath),
-			ObjectID: objectID,
-		}, nil)
-		w.repo.mod.Objects.Receive(&objects.EventDiscovered{
-			ObjectID: objectID,
-			Zone:     astral.ZoneDevice,
-		}, nil)
-
-	case strings.Contains(err.Error(), "UNIQUE constraint failed"):
-		err = objects.ErrAlreadyExists
-	}
-
-	if objectID != nil {
-		w.repo.pushAdded(objectID)
-	}
-
-	return objectID, err
+	return objectID, nil
 }
 
 func (w *Writer) Discard() error {
